create debit and withdraw flux

+++ app/src/main/resources/db/V1__CREATE_CUSTOMERS.sql
@@ -3,7 +3,7 @@ CREATE TABLE `customers`
  `document` varchar(13) NOT NULL ,
  `name`     varchar(100) NOT NULL ,
  `creation` datetime NOT NULL ,
- `balance`  decimal NOT NULL ,
+ `balance`  float NOT NULL ,
 
 PRIMARY KEY (`document`)
 );
